Move pkg/storage/sql to pkg/storage
