+++ garden-app/pkg/storage/db/weather_client_queries.sql.go
