+++ garden-app/pkg/storage/queries/weather_client_queries.sql
