+++ garden-app/pkg/storage/queries/zone_queries.sql
