+++ garden-app/pkg/storage/zones.go
@@ -1,4 +1,4 @@
-package sql
+package storage
 
 import (
 	"context"
@@ -11,7 +11,7 @@ import (
 	"time"
 
 	"github.com/calvinmclean/automated-garden/garden-app/pkg"
-	"github.com/calvinmclean/automated-garden/garden-app/pkg/storage/sql/db"
+	"github.com/calvinmclean/automated-garden/garden-app/pkg/storage/db"
 	"github.com/calvinmclean/babyapi"
 	"github.com/rs/xid"
 )
